chater/requirements.txt
black==23.3.0
cryptography==40.0.1
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
isort==5.12.0
mysqlclient==2.1.1
openai==0.27.4
PyJWT==2.6.0
redis==4.5.4
tiktoken==0.3.3